third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-token-at-position-expected.txt

Test editor tokenAtTextPosition method.


Running: testHighlightedText
Line: function foo(a, b) {
Column #0 (char 'f') - token: {"startColumn":0,"endColumn":8,"type":"js-keyword"}
Column #10 (char 'o') - token: {"startColumn":9,"endColumn":12,"type":"js-def"}
Column #13 (char 'a') - token: {"startColumn":13,"endColumn":14,"type":"js-def"}
Line:     var f = /.a/.test(a); /*
Column #8 (char 'f') - token: {"startColumn":8,"endColumn":9,"type":"js-def"}
Column #14 (char 'a') - token: {"startColumn":12,"endColumn":16,"type":"js-string-2"}
Column #19 (char 's') - token: {"startColumn":17,"endColumn":21,"type":"js-property"}
Line: this is a comment */
Column #0 (char 't') - token: {"startColumn":0,"endColumn":20,"type":"js-comment"}
Column #5 (char 'i') - token: {"startColumn":0,"endColumn":20,"type":"js-comment"}
Line:     return f + "looongword";
Column #5 (char 'e') - token: {"startColumn":4,"endColumn":10,"type":"js-keyword"}
Column #20 (char 'n') - token: {"startColumn":15,"endColumn":27,"type":"js-string"}
